chore(tools): move background-task execution functions to rara-background-tasks crate
