+++ src/tools/bash.rs
@@ -639,12 +639,16 @@ impl Tool for BashTool {
             .ok_or_else(|| ToolError::ExecutionFailed("stderr pipe unavailable".into()))?;
 
         let (tx, mut rx) = mpsc::channel(64);
-        let stdout_task = tokio::spawn(read_stream_chunks(
+        let stdout_task = tokio::spawn(rara_background_tasks::read_stream_chunks(
             stdout,
             BashStreamKind::Stdout,
             tx.clone(),
         ));
-        let stderr_task = tokio::spawn(read_stream_chunks(stderr, BashStreamKind::Stderr, tx));
+        let stderr_task = tokio::spawn(rara_background_tasks::read_stream_chunks(
+            stderr,
+            BashStreamKind::Stderr,
+            tx,
+        ));
 
         let mut stdout_text = String::new();
         let mut stderr_text = String::new();
@@ -713,7 +717,7 @@ impl Tool for BashTool {
                             stream: ToolOutputStream::Stderr,
                             chunk,
                         });
-                        kill_child_process_group(process_group_id);
+                        rara_background_tasks::kill_child_process_group(process_group_id);
                         let _ = child.start_kill();
                         cancellation_watchdog
                             .as_mut()
@@ -789,14 +793,21 @@ fn spawn_background_bash_task(
     stop_rx: oneshot::Receiver<()>,
     process_group_id: Option<u32>,
 ) {
+    let sandbox_warning =
+        if !wrapped.sandboxed && sandbox_permissions != BashSandboxPermissions::RequireEscalated {
+            Some(unsandboxed_execution_warning(&wrapped))
+        } else {
+            None
+        };
+    let cleanup_path = wrapped.cleanup_path.clone();
     tokio::spawn(async move {
-        let result = run_background_bash_task(
+        let result = rara_background_tasks::run_background_bash_task(
             &mut child,
-            wrapped,
-            sandbox_permissions,
             &record,
             stop_rx,
             process_group_id,
+            sandbox_warning.as_deref(),
+            cleanup_path.as_deref(),
         )
         .await;
         let (status, exit_code) = match result {
@@ -819,110 +830,6 @@ fn spawn_background_bash_task(
         };
         store.finish(&record.id, status, exit_code);
     });
-}
-
-async fn run_background_bash_task(
-    child: &mut Child,
-    wrapped: WrappedCommand,
-    sandbox_permissions: BashSandboxPermissions,
-    record: &BackgroundTaskRecord,
-    mut stop_rx: oneshot::Receiver<()>,
-    process_group_id: Option<u32>,
-) -> Result<Option<i32>, ToolError> {
-    if let Some(parent) = record.output_path.parent() {
-        fs::create_dir_all(parent).await?;
-    }
-    fs::write(&record.output_path, "").await?;
-    if !wrapped.sandboxed && sandbox_permissions != BashSandboxPermissions::RequireEscalated {
-        append_background_output(
-            &record.output_path,
-            BashStreamKind::Stderr,
-            &unsandboxed_execution_warning(&wrapped),
-        )
-        .await?;
-    }
-
-    let stdout = child
-        .stdout
-        .take()
-        .ok_or_else(|| ToolError::ExecutionFailed("stdout pipe unavailable".into()))?;
-    let stderr = child
-        .stderr
-        .take()
-        .ok_or_else(|| ToolError::ExecutionFailed("stderr pipe unavailable".into()))?;
-    let (tx, mut rx) = mpsc::channel(64);
-    let stdout_task = tokio::spawn(read_stream_chunks(
-        stdout,
-        BashStreamKind::Stdout,
-        tx.clone(),
-    ));
-    let stderr_task = tokio::spawn(read_stream_chunks(stderr, BashStreamKind::Stderr, tx));
-
-    let mut output_file = fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(&record.output_path)
-        .await?;
-    let mut stop_requested = false;
-    loop {
-        tokio::select! {
-            chunk = rx.recv() => {
-                let Some((stream, chunk)) = chunk else {
-                    break;
-                };
-                if !chunk.is_empty() {
-                    match stream {
-                        BashStreamKind::Stdout => output_file.write_all(chunk.as_bytes()).await?,
-                        BashStreamKind::Stderr => {
-                            output_file.write_all(b"[stderr] ").await?;
-                            output_file.write_all(chunk.as_bytes()).await?;
-                        }
-                    }
-                }
-            }
-            _ = &mut stop_rx, if !stop_requested => {
-                stop_requested = true;
-                kill_child_process_group(process_group_id);
-                let _ = child.start_kill();
-                output_file.write_all(b"[stderr] background task stop requested\n").await?;
-            }
-        }
-    }
-
-    let status = child.wait().await?;
-    stdout_task
-        .await
-        .map_err(|err| ToolError::ExecutionFailed(err.to_string()))??;
-    stderr_task
-        .await
-        .map_err(|err| ToolError::ExecutionFailed(err.to_string()))??;
-    if let Some(path) = wrapped.cleanup_path.as_ref() {
-        let _ = fs::remove_file(path).await;
-    }
-    Ok(status.code())
-}
-
-async fn read_stream_chunks<R>(
-    reader: R,
-    stream: BashStreamKind,
-    tx: mpsc::Sender<(BashStreamKind, String)>,
-) -> Result<(), ToolError>
-where
-    R: AsyncRead + Unpin + Send + 'static,
-{
-    let mut reader = reader;
-    let mut buffer = [0_u8; 4096];
-    loop {
-        let read = reader.read(&mut buffer).await?;
-        if read == 0 {
-            break;
-        }
-        let chunk = String::from_utf8_lossy(&buffer[..read]).into_owned();
-        if tx.send((stream, chunk)).await.is_err() {
-            break;
-        }
-    }
-    Ok(())
 }
 
 fn append_aggregated_bash_output(
@@ -985,25 +892,6 @@ fn configure_process_group(command: &mut Command) {
     #[cfg(not(unix))]
     {
         let _ = command;
-    }
-}
-
-fn kill_child_process_group(child_id: Option<u32>) {
-    #[cfg(unix)]
-    {
-        let Some(child_id) = child_id else {
-            return;
-        };
-        let process_group = -(child_id as libc::pid_t);
-        // Best-effort cancellation: the direct child may have already exited,
-        // but killing the group stops shell descendants that still hold pipes.
-        unsafe {
-            libc::kill(process_group, libc::SIGKILL);
-        }
-    }
-    #[cfg(not(unix))]
-    {
-        let _ = child_id;
     }
 }
